Generates the data. Finds some matches. Need better meserds
Finds contours of characters adequately. Doesn't do a great job in
translating to text. Could possibly be fixed from different
preprocessing or bigger training library. It would also be beneficial to
come up with a better way to train the data. Right now you have to enter
the capital character that program senses for each contour.